+++ join_impl/src/chain/expr/err_expr.rs
@@ -8,7 +8,7 @@ use syn::Expr;
 use super::InnerExpr;
 
 ///
-/// `ErrExpr` used to define types of expression which will be evaluated in case of `Err` or `None`.
+/// Used to define types of expression which will be evaluated in case of `Err` or `None`.
 ///
 #[derive(Clone, PartialEq, Eq, Debug)]
 pub enum ErrExpr {
@@ -50,7 +50,7 @@ impl ToTokens for ErrExpr {
 }
 
 impl InnerExpr for ErrExpr {
-    fn extract_inner(&self) -> Option<&[Expr]> {
+    fn get_inner_exprs(&self) -> Option<&[Expr]> {
         Some(match self {
             Self::Or(expr) => expr,
             Self::OrElse(expr) => expr,
@@ -58,8 +58,8 @@ impl InnerExpr for ErrExpr {
         })
     }
 
-    fn replace_inner(self, exprs: &[Expr]) -> Option<Self> {
-        exprs.last().map(Clone::clone).map(|expr| match self {
+    fn replace_inner_exprs(self, exprs: &[Expr]) -> Option<Self> {
+        exprs.last().cloned().map(|expr| match self {
             Self::Or(_) => Self::Or([expr]),
             Self::OrElse(_) => Self::OrElse([expr]),
             Self::MapErr(_) => Self::MapErr([expr]),
@@ -87,7 +87,10 @@ mod tests {
         ]
         .into_iter()
         {
-            assert_eq!(err_expr.extract_inner().clone(), Some(&[expr.clone()][..]));
+            assert_eq!(
+                err_expr.get_inner_exprs().clone(),
+                Some(&[expr.clone()][..])
+            );
         }
     }
 
@@ -105,9 +108,9 @@ mod tests {
         {
             assert_eq!(
                 err_expr
-                    .replace_inner(&[replace_inner.clone()][..])
+                    .replace_inner_exprs(&[replace_inner.clone()][..])
                     .unwrap()
-                    .extract_inner()
+                    .get_inner_exprs()
                     .clone(),
                 Some(&[replace_inner.clone()][..])
             );